+++ CHANGELOG.md
@@ -8,6 +8,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.7.4 - 2022-12-26
+
+### Added
+* Fetch fields based on current operation
+
+### Fixed
+* Problems with v-init inside FormBuilder
+* Refresh on edit page
+
+
 ## 0.7.3 - 2022-12-23
 
 ### Added